tinyCI single binary
This consolidates all services under a single binary. Please note that
this edition still expects each service to be launched indepedently;
more work needs to be done on that front.

Each service is listed under `tinyci service` for launching. Make tasks
should function normally as they have been migrated.
