+++ errors/error.go
@@ -162,7 +162,7 @@ func (e *Error) Contains(err interface{}) bool {
 
 // Exit exits the program leveraging the error for output before exiting with
 // error code 1. If DEBUG is set, it will output a stack trace.
-func (e Error) Exit() {
+func (e *Error) Exit() {
 	if os.Getenv("DEBUG") != "" {
 		fmt.Fprintf(os.Stderr, "%+v\n", e)
 	} else {
